chore: move common test
